Completely untested new relay board code.
This version supports three relays and does not read manual switches,
but instead indicator contacts from the relays. Therefore, it acts only
as a serial communication channel and does not arbitrate between inputs;
the assumption is that manual switches (if any) will be wired to
be ORed with the Arduino outputs.